massa nytt

+++ Styrenheten/StyrReglering.h
@@ -19,7 +19,7 @@ void turnLeft90();
 void regulateStraight();
 void setSpeedRight(uint8_t speed);
 void setSpeedLeft(uint8_t speed);
-int8_t degToRad(int8_t degree);
+int16_t degToRad(int8_t degree);
 void setDirLeft(uint8_t dir);
 void setDirRight(uint8_t dir);
 int8_t getRelativeY(void);
